requirements.txt
flask
motor
python-dotenv
pydantic
